restructure directory (#29)
