+++ src/autogluon_dashboard/constants/plots_constants.py
