+++ src/autogluon_dashboard/constants/widgets_constants.py
@@ -1,4 +1,9 @@
-from .df_constants import BESTDIFF, LOSS_RESCALED, TIME_INFER_S_RESCALED, TIME_TRAIN_S_RESCALED
+from autogluon_dashboard.constants.df_constants import (
+    BESTDIFF,
+    LOSS_RESCALED,
+    TIME_INFER_S_RESCALED,
+    TIME_TRAIN_S_RESCALED,
+)
 
 METRICS_TO_PLOT = [
     LOSS_RESCALED,
