+++ src/autogluon_dashboard/plotting/metric_counts_framework.py
@@ -3,7 +3,7 @@ from typing import List, Optional, Union
 import hvplot
 import pandas
 
-from ..scripts.utils import get_col_metric_counts, get_df_filter_by_framework
+from ..utils.dataset_utils import get_col_metric_counts, get_df_filter_by_framework
 from .all_plots import Plot
