+++ src/autogluon_dashboard/plotting/top5_per_dataset.py
@@ -3,7 +3,7 @@ from typing import List, Optional, Union
 import hvplot
 import pandas
 
-from ..scripts.utils import get_df_filter_by_dataset, get_top5_performers
+from ..utils.dataset_utils import get_df_filter_by_dataset, get_top5_performers
 from .all_plots import Plot
